+++ modules/network/variables.tf
@@ -1,15 +1,10 @@
-## Copyright © 2019, Oracle and/or its affiliates. 
+## Copyright (c) 2020, Oracle and/or its affiliates.
 ## All rights reserved. The Universal Permissive License (UPL), Version 1.0 as shown at http://oss.oracle.com/licenses/upl
 
 # OCI Service
 variable "compartment_id" {
   description = "Compartment OCID where the VCN is created."
   type        = string
-}
-
-variable "use_existing_vcn" {
-  type    = bool
-  default = false
 }
 
 variable "existing_vcn_id" {
@@ -42,3 +37,13 @@ variable "enable_nat_gateway" {
   default = false
 }
 
+// Tags
+variable "freeform_tags" {
+  type = map(string)
+  default = null
+}
+
+variable "defined_tags" {
+  type = map(string)
+  default = null
+}