+++ modules/subscription/variables.tf
@@ -1,8 +1,13 @@
-## Copyright © 2019, Oracle and/or its affiliates. 
+## Copyright (c) 2020, Oracle and/or its affiliates.
 ## All rights reserved. The Universal Permissive License (UPL), Version 1.0 as shown at http://oss.oracle.com/licenses/upl
 
 variable "compartment_id" {
   type = string
+}
+
+variable "enabled" {
+  type = bool
+  default = false
 }
 
 variable "mp_listing_id" {